2008-09-08  Doug Kwan  <dougkwan@google.com>
	* lto-plugin.c (all_symbols_read_handler): Widen fprintf arguments
	to 64-bit to fix a 32-bit bootstrap failure due to format warnings.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/lto@140113 138bc75d-0d04-0410-961f-82ee72b054a4

+++ lto-plugin/lto-plugin.c
@@ -277,8 +277,8 @@ all_symbols_read_handler (void)
 	  uint32_t slot = symtab->slots[j];
 	  unsigned int resolution = syms[j].resolution;
 	  fprintf(f, "%s\n%" PRId64 " %" PRId64 " %d %s\n",
-		  file->name, file->offset, file->filesize, slot,
-		  lto_resolution_str[resolution]);
+		  file->name, (int64_t) file->offset, (int64_t) file->filesize,
+		  slot, lto_resolution_str[resolution]);
 	}
       free (syms);
     }
